@@ -1,9 +1,9 @@
-const Hapi = require('@hapi/hapi');
-const {routes} = require('./routes');
+import Hapi from '@hapi/hapi';
+import routes from './routes.js';
 
 const init = async () => {
   const server = Hapi.server({
-    port: 5000,
+    port: 9000,
     host: 'localhost',
     routes: {
       cors: {
