@@ -1,4 +1,4 @@
 const books = [];
 const book = [];
 
-module.exports = {books, book};
+export {books, book};
